requirements.txt
pymongo
requests
apache-airflow
streamlit
pandas
plotly
matplotlib
